+++ vm_sources/commands/proc_09_zjmp.c
@@ -23,7 +23,7 @@ void			proc_zjump(void *data)
 	//ft_printf("field %#x\n", get_field_val(proc->pc));
 	//ft_printf("gonna jump. carry %d\n", proc->carry);
 	if (proc->carry)
-		proc->pc = set_pos(proc->pc + (get_int(proc->pc + 1, 2)) % IDX_MOD);
+		proc->pc = set_pos(proc->pc + (short)((get_int(proc->pc + 1, 2))) % IDX_MOD);
 	else
 		proc->pc = set_pos(proc->pc + 3);
 }
