+++ vm_sources/commands/proc_10_ldi.c
@@ -43,7 +43,7 @@ void			proc_ldi(void *data)
 	if (cod_b.t[0] == TREG)
 		arg[0] = proc->registry[arg[0]];
 	else if (cod_b.t[0] == TIND)
-		arg[0] = get_int(proc->pc + (arg[0]) % IDX_MOD, 4);
+		arg[0] = get_int(proc->pc + (short)(arg[0]) % IDX_MOD, 4);
 
 	if (cod_b.t[1] == TREG)
 		arg[1] = proc->registry[arg[1]];
