+++ vm_sources/commands/proc_11_sti.c
@@ -46,7 +46,7 @@ void			proc_sti(void *data)
 	if (cod_b.t[1] == TREG)
 		arg[1] = proc->registry[arg[1]];
 	else if (cod_b.t[1] == TIND)
-			arg[1] = get_int(proc->pc + (arg[1]) % IDX_MOD, 4);
+			arg[1] = get_int(proc->pc + (short)(arg[1]) % IDX_MOD, 4);
 
 	if (cod_b.t[2] == TREG)
 		arg[2] = proc->registry[arg[2]];
